confirmación al eliminar un amigo

+++ frontend/chessandconnect/src/app/components/friends-list/friends-list.component.ts
@@ -3,6 +3,7 @@ import { FriendsService } from '../../services/friends.service';
 import { Friend } from '../../models/dto/friend';
 import { User } from '../../models/dto/user';
 import { RouterLink } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-friends-list',
@@ -23,10 +24,33 @@ export class FriendsListComponent implements OnInit {
   }
   
 
-  async deleteFriend(friendId: number){
-    console.log("Eliminado" + friendId)
-    await this.friendService.deleteFriend(friendId)
+
+async deleteFriend(friendId: number) {
+  //Mensaje de confirmación
+  const result = await Swal.fire({
+    title: '¿Estás seguro?',
+    text: 'No podrás revertir esta acción.',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Sí, eliminar',
+    cancelButtonText: 'Cancelar',
+    customClass: {
+      popup: 'bg-[#301e16] text-[#E8D5B5]',  
+      title: 'font-bold text-lg',  
+      confirmButton: 'bg-[#CBA77B] hover:bg-[#A68556] text-[#301e16] font-medium py-2 px-4 rounded-lg',  
+      cancelButton: 'bg-[#CBA77B] hover:bg-[#A68556] text-[#301e16] font-medium py-2 px-4 rounded-lg'  
+    }
+  });
+
+  if (result.isConfirmed) {
+    console.log("Amigo eliminado: " + friendId);
+    await this.friendService.deleteFriend(friendId);
+  } else {
+    console.log("Eliminación cancelada");
   }
+}
+
+
